Created cli.py as entry point for the CLI usage of amami. Now __init__.py only contains the main dunder methods for amami (__docs__, __version__, etc.). Enabled '--poor' options that disables 'rich' package functionality for terminal styling/colouring, and renders raw output.
