+++ setup.cfg
@@ -33,4 +33,4 @@ install_requires =
 
 [entry_points]
 console_scripts = 
-    amami=amami:main
+    amami=amami.cli:main
